Removed roles file
